requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo
python-dotenv
APScheduler
ollama
beautifulsoup4
numpy
torch
transformers
pydub
python-ffmpeg
pytz